optimization & readability: fix style & organization
