--- src/app/components/analyse/sequencer/create-label-btn-dialog.component.scss
@@ -1,39 +1,0 @@
-.dialog-title {
-  margin: 0;
-}
-
-.dialog-body {
-  padding-top: 12px;
-}
-
-.dialog-actions {
-  padding-bottom: 12px;
-}
-
-.dialog-form {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.row {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(160px, 1fr));
-  gap: 12px;
-}
-
-.radio-group {
-  display: flex;
-  gap: 16px;
-}
-
-.section h4 {
-  margin: 0 0 8px;
-  font-size: 13px;
-  font-weight: 600;
-}
-
-.error {
-  font-size: 12px;
-  margin-top: 6px;
-}
