+++ src/app/components/analyse/sequencer/createBtn/event/create-event-btn-dialog.component.html
@@ -1,9 +1,14 @@
-<h2 mat-dialog-title class="bg-layer-3 light-text dialog-title">
+<h2 mat-dialog-title class="bg-layer-3 light-text mb-3 text-center">
   {{ isEdit ? 'Éditer un Event' : 'Créer un Event' }}
 </h2>
 
 <mat-dialog-content class="bg-layer-3 light-text dialog-body">
-  <form [formGroup]="form" class="dialog-form">
+  <form [formGroup]="form" class="flex flex-col gap-4">
+    <mat-radio-group class="flex justify-center gap-2" formControlName="kind">
+      <mat-radio-button value="limited">Limited</mat-radio-button>
+      <mat-radio-button value="indefinite">Indefinite</mat-radio-button>
+    </mat-radio-group>
+
     <div class="row">
       <mat-form-field appearance="fill">
         <mat-label>ID</mat-label>
@@ -14,11 +19,6 @@
         <input matInput formControlName="name" />
       </mat-form-field>
     </div>
-
-    <mat-radio-group class="radio-group" formControlName="kind">
-      <mat-radio-button value="limited">Limited</mat-radio-button>
-      <mat-radio-button value="indefinite">Indefinite</mat-radio-button>
-    </mat-radio-group>
 
     <div class="row">
       <mat-form-field appearance="fill">
@@ -45,7 +45,7 @@
   </form>
 </mat-dialog-content>
 
-<mat-dialog-actions align="end" class="bg-layer-3 light-text dialog-actions">
+<mat-dialog-actions align="center" class="bg-layer-3 light-text pb-3">
   <button mat-stroked-button type="button" (click)="close()">Annuler</button>
   <button mat-flat-button color="primary" type="button" (click)="save()" [disabled]="!canSave()">
     {{ isEdit ? 'Mettre à jour' : 'Créer' }}
