+++ src/app/components/analyse/sequencer/createBtn/label/create-label-btn-dialog.component.scss
@@ -1,0 +1,16 @@
+.row {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(160px, 1fr));
+  gap: 12px;
+}
+
+.section h4 {
+  margin: 0 0 8px;
+  font-size: 13px;
+  font-weight: 600;
+}
+
+.error {
+  font-size: 12px;
+  margin-top: 6px;
+}
