+++ src/app/components/analyse/sequencer/hotkeyPicker/hotkey-picker.component.html
