+++ src/app/components/analyse/sequencer/hotkeyPicker/hotkey-picker.component.scss
