+++ src/theme/_global.class.scss
@@ -22,11 +22,11 @@
 }
 
 .light-text {
-  color: var(--light-text);
+  color: var(--light);
 }
 
 .dark-text {
-  color: var(--dark-text);
+  color: var(--dark);
 }
 
 .primary-text {
