+++ src/theme/variables/_global.variables.scss
@@ -5,9 +5,9 @@
   --bg-secondary-layer-1: #FFC107;
   --bg-secondary-layer-2: #FF5722;
 
-  --light-text: #FAFAFA;
+  --light: #FAFAFA;
   --light-text-disabled: #C3C3C3;
-  --dark-text: #OCOCOC;
+  --dark: #0c0c0c;
   --primary-text: #278BC9;
 }
 
@@ -17,8 +17,8 @@
   --bg-secondary-layer-1: #FFC107;
   --bg-secondary-layer-2: #FF5722;
 
-  --light-text: #FAFAFA;
+  --light: #FAFAFA;
   --light-text-disabled: #C3C3C3;
-  --dark-text: #0C0C0C;
+  --dark: #0C0C0C;
   --primary-text: #278BC9;
 }
